Move BuildInfo class to own file
