+++ src/main.ts
@@ -4,7 +4,8 @@ import * as fs from 'fs'
 
 import {Commander} from './commander'
 import {LaTeXCommander} from './components/commander'
-import {Logger, BuildInfo} from './components/logger'
+import {Logger} from './components/logger'
+import {BuildInfo} from './components/buildinfo'
 import {Manager} from './components/manager'
 import {Builder} from './components/builder'
 import {Viewer} from './components/viewer'
